+++ WDE.SmartScriptEditor.WPF/Editor/Views/Helpers/ParameterDataTemplateSelector.cs
@@ -14,7 +14,7 @@ namespace WDE.SmartScriptEditor.WPF.Editor.Views.Helpers
         
         public override DataTemplate SelectTemplate(object item, DependencyObject container)
         {
-            if (item is EditableParameterViewModel<int> intParam && intParam.Parameter.Parameter is BoolParameter boolParameter)
+            if (item is EditableParameterViewModel<long> intParam && intParam.Parameter.Parameter is BoolParameter boolParameter)
                 return BoolParameter;
             if (item is EditableParameterActionViewModel aevm)
                 return ButtonParameter;
