+++ WDE.SmartScriptEditor/Editor/ViewModels/SmartScriptEditorViewModel.cs
@@ -23,6 +23,7 @@ using WDE.Conditions.Data;
 using WDE.Conditions.Exporter;
 using WDE.MVVM;
 using WDE.MVVM.Observable;
+using WDE.Parameters.Models;
 using WDE.SmartScriptEditor.Data;
 using WDE.SmartScriptEditor.Editor.UserControls;
 using WDE.SmartScriptEditor.Editor.ViewModels.Editing;
@@ -231,9 +232,9 @@ namespace WDE.SmartScriptEditor.Editor.ViewModels
 
             DirectEditParameter = new DelegateCommand<object>(async obj =>
             {
-                if (obj is ParameterValueHolder<int> param)
+                if (obj is ParameterValueHolder<long> param)
                 {
-                    int? val = await itemFromListProvider.GetItemFromList(param.Parameter.Items ?? new Dictionary<int, SelectOption>(), param.Parameter is FlagParameter, param.Value);
+                    long? val = await itemFromListProvider.GetItemFromList(param.Parameter.Items ?? new Dictionary<long, SelectOption>(), param.Parameter is FlagParameter, param.Value);
                     if (val.HasValue)
                         param.Value = val.Value;   
                 } 
@@ -887,8 +888,8 @@ namespace WDE.SmartScriptEditor.Editor.ViewModels
 
         private async Task AsyncLoad()
         {
-            var lines = database.GetScriptFor(this.item.Entry, this.item.SmartType).ToList();
-            var conditions = database.GetConditionsFor(SmartConstants.ConditionSourceSmartScript, this.item.Entry, (int)this.item.SmartType).ToList();
+            var lines = database.GetScriptFor((int)this.item.Entry, this.item.SmartType).ToList();
+            var conditions = database.GetConditionsFor(SmartConstants.ConditionSourceSmartScript, (int)this.item.Entry, (int)this.item.SmartType).ToList();
             script.Load(lines, conditions);
             IsLoading = false;
             History.AddHandler(new SaiHistoryHandler(script, smartFactory));
@@ -900,11 +901,11 @@ namespace WDE.SmartScriptEditor.Editor.ViewModels
 
             var (lines, conditions) = script.ToSmartScriptLinesNoMetaActions(smartFactory, smartDataManager);
             
-            await database.InstallScriptFor(item.Entry, item.SmartType, lines);
+            await database.InstallScriptFor((int)item.Entry, item.SmartType, lines);
             
             await database.InstallConditions(conditions, 
                 IDatabaseProvider.ConditionKeyMask.SourceEntry | IDatabaseProvider.ConditionKeyMask.SourceId,
-                new IDatabaseProvider.ConditionKey(SmartConstants.ConditionSourceSmartScript, null, item.Entry, (int)item.SmartType));
+                new IDatabaseProvider.ConditionKey(SmartConstants.ConditionSourceSmartScript, null, (int)item.Entry, (int)item.SmartType));
 
             statusbar.PublishNotification(new PlainNotification(NotificationType.Success, "Saved to database"));
             
@@ -1052,7 +1053,7 @@ namespace WDE.SmartScriptEditor.Editor.ViewModels
                 obj.Parent = originalAction.Parent; // fake parent
             SmartGenericJsonData actionData = smartDataManager.GetRawData(SmartType.SmartAction, originalAction.Id);
             
-            var parametersList = new List<(ParameterValueHolder<int>, string)>();
+            var parametersList = new List<(ParameterValueHolder<long>, string)>();
             var floatParametersList = new List<(ParameterValueHolder<float>, string)>();
             var actionList = new List<EditableActionData>();
             var stringParametersList = new List<(ParameterValueHolder<string>, string)>() {(obj.CommentParameter, "Comment")};
@@ -1174,7 +1175,7 @@ namespace WDE.SmartScriptEditor.Editor.ViewModels
                 obj.Parent = originalCondition.Parent; // fake parent
             
             var actionList = new List<EditableActionData>();
-            var parametersList = new List<(ParameterValueHolder<int>, string)>();
+            var parametersList = new List<(ParameterValueHolder<long>, string)>();
 
             parametersList.Add((obj.Inverted, "General"));
             parametersList.Add((obj.ConditionTarget, "General"));
@@ -1238,7 +1239,7 @@ namespace WDE.SmartScriptEditor.Editor.ViewModels
                 ev.Parent = originalEvent.Parent; // fake parent
             
             var actionList = new List<EditableActionData>();
-            var parametersList = new List<(ParameterValueHolder<int>, string)>
+            var parametersList = new List<(ParameterValueHolder<long>, string)>
             {
                 (ev.Chance, "General"),
                 (ev.Flags, "General"),
