+++ WDE.SmartScriptEditor/History/SaiHistoryHandler.cs
@@ -2,6 +2,7 @@
 using System.Collections.Specialized;
 using WDE.Common.History;
 using WDE.MVVM;
+using WDE.Parameters.Models;
 using WDE.SmartScriptEditor.Data;
 using WDE.SmartScriptEditor.Models;
 using WDE.SmartScriptEditor.Utils;
@@ -265,7 +266,7 @@ namespace WDE.SmartScriptEditor.History
             }
         }
 
-        private void Parameter_OnValueChanged(ParameterValueHolder<int> sender, int old, int @new)
+        private void Parameter_OnValueChanged(ParameterValueHolder<long> sender, long old, long @new)
         {
             PushAction(new ParameterChangedAction(sender, old, @new));
         }
@@ -369,9 +370,9 @@ namespace WDE.SmartScriptEditor.History
             }
         }
 
-        private class ParameterChangedAction : GenericParameterChangedAction<int>
+        private class ParameterChangedAction : GenericParameterChangedAction<long>
         {
-            public ParameterChangedAction(ParameterValueHolder<int> param, int old, int @new) : base(param, old, @new)
+            public ParameterChangedAction(ParameterValueHolder<long> param, long old, long @new) : base(param, old, @new)
             {
             }
         }
